+++ src/main/java/com/atlassian/db/replica/api/DualConnection.java
@@ -104,8 +104,13 @@ public final class DualConnection implements Connection {
     @Override
     public String nativeSQL(String sql) throws SQLException {
         checkClosed();
-        return connectionProvider.getWriteConnection(new RouteDecisionBuilder(Reason.RW_API_CALL).sql(sql)).nativeSQL(
-            sql);
+        if (compatibleWithPreviousVersion) {
+            return connectionProvider.getWriteConnection(new RouteDecisionBuilder(Reason.RW_API_CALL).sql(sql))
+                .nativeSQL(sql);
+        } else {
+            return connectionProvider.getReadConnection(new RouteDecisionBuilder(Reason.RO_API_CALL).sql(sql))
+                .nativeSQL(sql);
+        }
     }
 
     @Override