+++ src/test/java/com/atlassian/db/replica/api/TestDualConnection.java
@@ -190,12 +190,27 @@ public class TestDualConnection {
     }
 
     @Test
-    public void shouldRunNativeSqlOnMain() throws SQLException {
+    public void shouldRunNativeSqlOnReplica() throws SQLException {
         final ConnectionProviderMock connectionProvider = new ConnectionProviderMock();
         final Connection connection = DualConnection.builder(
             connectionProvider,
-            permanentInconsistency().build()
+            permanentConsistency().build()
         ).build();
+
+        connection.nativeSQL(SIMPLE_QUERY);
+
+        assertThat(connectionProvider.getProvidedConnectionTypes())
+            .containsExactly(REPLICA);
+    }
+
+    @Test
+    public void shouldBePossibleToDisableNativeSqlFix() throws SQLException {
+        final ConnectionProviderMock connectionProvider = new ConnectionProviderMock();
+        final Connection connection = DualConnection.builder(
+            connectionProvider,
+            permanentConsistency().build()
+        ).compatibleWithPreviousVersion()
+            .build();
 
         connection.nativeSQL(SIMPLE_QUERY);